Move JVM options to the right location
Currently we put our custom JVM options into the `/etc/default/tomcat7`.
That's obviously wrong, because we don't start Tomcat via SysVinit, but
directly with `catalina.sh`. So this commits adds the respective
environment variable to the call to `catalina.sh` and removes the change
of `JAVA_OPTS` in `/etc/default/tomcat7`.