Use new way of deployment for OSIAM
OSIAM now loads assets and configuration from a directory in the
file system known as `OSIAM_HOME`.
Add a Tomcat deployment descriptor to set the home directory for the
deployment.
Put the home directory under `/var/lib/osiam`, because it fits quite
nicely there.
This is also the home directory used in the official docs.